refactor open file
