+++ pkg/deej/util/util_linux.go
@@ -2,26 +2,13 @@ package util
 
 import (
 	"errors"
-	"fmt"
 	"os/exec"
-
-	"go.uber.org/zap"
 )
 
 func getCurrentWindowProcessNames() ([]string, error) {
-	return nil, errors.New("Not implemented")
+	return nil, errors.New("not implemented")
 }
 
-func OpenExternal(logger *zap.SugaredLogger, filename string) error {
-	command := exec.Command("xdg-open", filename)
-
-	if err := command.Run(); err != nil {
-		logger.Warnw("Failed to open file",
-			"filename", filename,
-			"error", err)
-
-		return fmt.Errorf("open file proc: %w", err)
-	}
-
-	return nil
+func getOpenExternalCommand(filename string) *exec.Cmd {
+	return exec.Command("xdg-open", filename)
 }
