+++ pkg/deej/util/util_windows.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/lxn/win"
 	"github.com/mitchellh/go-ps"
-	"go.uber.org/zap"
 )
 
 const (
@@ -99,16 +98,6 @@ func getCurrentWindowProcessNames() ([]string, error) {
 	return result, nil
 }
 
-func OpenExternal(logger *zap.SugaredLogger, filename string) error {
-	command := exec.Command(filepath.Join(os.Getenv("SYSTEMROOT"), "System32", "rundll32.exe"), "url.dll,FileProtocolHandler", filename)
-
-	if err := command.Start(); err != nil {
-		logger.Warnw("Failed to open file",
-			"filename", filename,
-			"error", err)
-
-		return fmt.Errorf("open file proc: %w", err)
-	}
-
-	return nil
+func getOpenExternalCommand(filename string) *exec.Cmd {
+	return exec.Command(filepath.Join(os.Getenv("SYSTEMROOT"), "System32", "rundll32.exe"), "url.dll,FileProtocolHandler", filename)
 }
